+++ src/py/pyodide/__init__.py
@@ -21,6 +21,14 @@ from _pyodide._base import (
 from _pyodide._importhook import register_js_module, unregister_js_module
 
 from . import _state  # noqa: F401
+from ._browser_apis import (
+    add_event_listener,
+    clear_interval,
+    clear_timeout,
+    remove_event_listener,
+    set_interval,
+    set_timeout,
+)
 from ._core import (
     IN_BROWSER,
     ConversionError,
@@ -62,4 +70,10 @@ __all__ = [
     "should_quiet",
     "ConversionError",
     "destroy_proxies",
+    "set_timeout",
+    "set_interval",
+    "clear_timeout",
+    "clear_interval",
+    "add_event_listener",
+    "remove_event_listener",
 ]